requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Data Science
pandas==2.2.0
numpy==1.26.3
xarray==2024.1.0
scipy==1.12.0
scikit-learn==1.4.0

# Meteorology
metpy==1.6.0
cfgrib==0.9.11.0
netCDF4==1.6.5

# Weather Data Access
siphon==0.9.0
requests==2.31.0

# Task Scheduling
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development Tools
ipython==8.20.0
jupyter==1.0.0
black==24.1.1
ruff==0.1.14

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
